+++ HISTORY.rst
@@ -31,6 +31,7 @@ TODO
 0.4.1.dev0
 ----------
 
+* Clearly state when creating samples (ignore samples if not defined in database)
 * Track sample aliases for ``original_id``
 * Import samples from file by providing *country* and *breeds* values as parameters
 * Security updates